removed warning.
git-svn-id: https://dune.mathematik.uni-freiburg.de/svn/dune-fem/trunk@4561 4a6c9bf5-c326-0410-8cc6-c7cabc8ddcf0

+++ fem/space/dgspace/dgadaptiveleafindexset.hh
@@ -526,7 +526,8 @@ public:
     }
 
     // for parallel data read in serial program we need compression 
-    if(StructuredGrid && (hIndexSet_.size(0) != codimLeafSet_.size()) )
+    if(StructuredGrid && 
+    	((int) hIndexSet_.size(0) != (int) codimLeafSet_.size()) )
     {
       compressed_ = false;
     }
